+++ include/nstd/events.h
@@ -25,77 +25,23 @@ typedef enum
     NSTD_EVENT_LOOP_CONTROL_FLOW_EXIT
 } NSTDEventLoopControlFlow;
 
-/// Represents an event.
-typedef enum
-{
-    /// There is no event.
-    NSTD_EVENT_LOOP_EVENT_NONE,
-    /// The event loop is about to be destroyed.
-    NSTD_EVENT_LOOP_EVENT_LOOP_DESTROYED,
-    /// All events have been cleared.
-    NSTD_EVENT_LOOP_EVENT_EVENTS_CLEARED,
-    /// A device has been added.
-    NSTD_EVENT_LOOP_EVENT_DEVICE_ADDED,
-    /// A device has been removed.
-    NSTD_EVENT_LOOP_EVENT_DEVICE_REMOVED,
-    /// The mouse has been moved.
-    NSTD_EVENT_LOOP_EVENT_MOUSE_MOVED,
-    /// The scroll wheel was scrolled.
-    NSTD_EVENT_LOOP_EVENT_SCROLL_PIXEL,
-    /// The scroll wheel was scrolled.
-    NSTD_EVENT_LOOP_EVENT_SCROLL_LINE,
-    /// A window requests a redraw.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_REDRAW_REQUESTED,
-    /// A window has been resized.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_RESIZED,
-    /// A window was moved.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_MOVED,
-    /// Window focus has changed.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_FOCUS_CHANGED,
-    /// A keyboard key was pressed.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_KEY,
-    /// A modifier key was pressed.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_MOD_KEY,
-    /// The mouse has moved.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_MOUSE_MOVED,
-    /// The mouse entered the window's frame.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_MOUSE_ENTERED,
-    /// The mouse left the window's frame.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_MOUSE_LEFT,
-    /// The scroll wheel was scrolled.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_SCROLL,
-    /// A mouse button was clicked.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_MOUSE_BUTTON,
-    /// A window requests closing.
-    NSTD_EVENT_LOOP_EVENT_WINDOW_CLOSE_REQUESTED
-} NSTDEventLoopEvent;
-
-/// Holds an event's data.
+/// Event callbacks.
 typedef struct
 {
-    /// The event that was recieved.
-    NSTDEventLoopEvent event;
-    /// The difference in mouse position.
-    NSTDFloat64 mouse_delta[2];
-    /// A size.
-    NSTDUInt32 size[2];
-    /// A position.
-    NSTDInt32 pos[2];
-    /// The ID of a window.
-    NSTDWindowID window_id;
-    /// Raw input.
-    NSTDRawInput raw_input;
-    /// Touch state.
-    NSTDTouchState touch_state;
-    /// The mouse button event.
-    NSTDMouseButtonEvent mouse_button_event;
-    /// The key.
-    NSTDKeyEvent key;
-    /// The modifier keys.
-    NSTDUInt8 mod_keys;
-    /// `NSTD_BOOL_TRUE` if the window has focus.
-    NSTDBool has_focus;
-} NSTDEventData;
+    /// Called when all main events have been processed.
+    /// Parameters:
+    ///     `NSTDEventLoopControlFlow *control_flow` - The control flow of the event loop.
+    void (*on_update)(NSTDEventLoopControlFlow *);
+    /// Called when a window requests closing.
+    /// Parameters:
+    ///     `NSTDEventLoopControlFlow *control_flow` - The control flow of the event loop.
+    ///     `NSTDWindowID window_id` - The ID of the window that requests closing.
+    void (*on_window_requests_closing)(NSTDEventLoopControlFlow *, NSTDWindowID);
+    /// Called when the event loop is being destroyed.
+    /// Parameters:
+    ///     `NSTDEventLoopControlFlow *control_flow` - The control flow of the event loop.
+    void (*on_destroy)(NSTDEventLoopControlFlow *);
+} NSTDEventCallbacks;
 
 /// Creates a new event loop.
 /// Returns: `NSTDEventLoop event_loop` - The event loop.
@@ -109,17 +55,21 @@ NSTDAPI NSTDEventLoop nstd_events_event_loop_new();
 ///     - Android
 /// Parameters:
 ///     `NSTDEventLoop *const event_loop` - The event loop to run.
-///     `NSTDEventLoopControlFlow(*callback)(NSTDEventData *)` - Called once per event.
+///     `const NSTDEventCallbacks *const callbacks` - The event callbacks.
 ///     `const NSTDBool should_return` - `NSTD_BOOL_TRUE` if this function should return.
 NSTDAPI void nstd_events_event_loop_run(
     NSTDEventLoop *const event_loop,
-    NSTDEventLoopControlFlow(*callback)(NSTDEventData *),
+    const NSTDEventCallbacks *const callbacks,
     const NSTDBool should_return);
 
 /// Frees an event loop without running it.
 /// Parameters:
 ///     `NSTDEventLoop *const event_loop` - The event loop to free.
 NSTDAPI void nstd_events_event_loop_free(NSTDEventLoop *const event_loop);
+
+/// Creates a new `NSTDEventCallbacks` with default callbacks.
+/// Returns: `NSTDEventCallbacks callbacks` - The default event callbacks.
+NSTDAPI NSTDEventCallbacks nstd_events_event_callbacks_default();
 
 #ifdef NSTDCPP
 }